+++ src/energy_systems/electric_producers/chpp.jl
@@ -67,7 +67,7 @@ function control(
     parameters::Dict{String,Any}
 )
     move_state(unit, components, parameters)
-    unit.output_interfaces[unit.m_heat_out].temperature = highest_temperature(unit.output_temperature, unit.output_interfaces[unit.m_heat_out].temperature)
+    unit.output_interfaces[unit.m_heat_out].temperature = highest(unit.output_temperature, unit.output_interfaces[unit.m_heat_out].temperature)
 end
 
 function set_max_energies!(unit::CHPP, gas_in::Float64, el_out::Float64, heat_out::Float64)