+++ src/energy_systems/heat_producers/fuel_boiler.jl
@@ -77,7 +77,7 @@ function control(
     parameters::Dict{String,Any}
 )
     move_state(unit, components, parameters)
-    unit.output_interfaces[unit.m_heat_out].temperature = highest_temperature(
+    unit.output_interfaces[unit.m_heat_out].temperature = highest(
         unit.output_temperature,
         unit.output_interfaces[unit.m_heat_out].temperature
     )